Added calibration for each TEMPer device on the system.
TemperDevice.__init__() reads /etc/temper.conf if it exists, and calibrates each device which has a configuration line in the file.
Format of the config file is, for example,

Device 0: scale = 0.937, offset = -0.5625

(see CONF_RE in temper.py for details).
Calibration is made with the formula

y = scale * x + offset

where x is the raw value read from the device (in celcius), y is the calibrated value, and scale and offset are read from the config file.

Since TemperDevice class object doesn't know what device number it is assigned,
an argument "id" is added to TemperDevice.get_temperature() to tell it which calibration data line to use.