Change the way TEMPer device identifies it's physical location.
TemperDevice now accepts "bus" argument in its __init__() method.
Initializer code looks upon usb sysfs directory using this bus and deviceID
information, and finds the port(or chain of ports) which this device is
plugged on. After that, it reads configuration file (/etc/temper.conf) and
uses calibration parameters found on the line with matching bus-ports.

This is needed to uniquely identify the physical location of a TEMPer device,
since only bus-ports combination can persist across reboots or connection
 losses.