Add basic support for the TEMPerHUM device
Added the TEMPerHUM USB device id to the list of known working devices
in the README.
Extended VIDPIDS list with the new values for the TEMPerHUM in
temper.py.

The device reports itself as TEMPer1F_H1_V1.4.
It is a single sensor device, but it reports both temperature (at offset
2-3) and relative humidity (at offset 4-5) from that sensor.